Fix bugs in MCMC implementation + unit 7 tweaks for code readability
The previous example MCMC implementation (MCMC via Metropolis-Hastings)
was actually quite buggy - when rejecting a step it didn't keep the
existing value of the state variable theta as the next state in the
markov chain, but kept retrying until acceptance.

This significantly underestimates the value of the posterior near the
maximum of the distribution (which you could see in the example graph).